Limits in users file are removed or gracefully handled.
While parsing 'users' file on server, BasicMail had two limits.

MAX_USERS imposed a maximum number of users. Parsing of file would
not check for a file too long, though, so buffer overflow was possible.
Now, BasicMail is able to parse an unlimited number of users.

MAX_FLEN imposes a maximum length for username and password fields.
This limitation is still present, but the application checks for illegal situations
while parsing file and avoids buffer overflows.
Depending on length of fields, an offending line may be accepted with one or
both fields truncated or may be skipped.